requirements.txt
paramiko>=2.4.2
astartool>=0.0.8
psutil
python-gitlab>=3.0.0